refactor label and message of inputbox
