--- packages/basic/demo/stories/FormGroup/Upload/docs.md
@@ -1,7 +1,0 @@
-# Upload
-
-### Usage
-
-```html
-<x-upload v-bind="scheme"/>
-```
